+++ packages/flame/CHANGELOG.md
@@ -38,6 +38,7 @@
  - Rename `HasTappableComponents` to `HasTappables`
  - Rename `HasDraggableComponents` to `HasDraggables`
  - Rename `HasHoverableComponents` to `HasHoverableis`
+ - Added `ScaleEffect` backed by the new effects engine
  - Update `OrderedSet` to 4.1.0
  - Update `OrderedSet` to 5.0.0